+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - Lib "Python-telegram-bot" has been updated to version 21
+ - Threads replaced by Asyncio
 
 ## [1.1.0] - 2023-12-24
 ### Added